Type test has been extracted from `CassandraConnectorIT` to its own IT suite.
